+++ xinference/model/embedding/core.py
@@ -24,8 +24,6 @@ from ...constants import XINFERENCE_CACHE_DIR
 from ...types import Embedding, EmbeddingData, EmbeddingUsage
 from ..core import ModelDescription
 
-MAX_ATTEMPTS = 3
-
 logger = logging.getLogger(__name__)
 
 
@@ -40,10 +38,10 @@ class EmbeddingModelSpec(BaseModel):
 
 def cache(model_spec: EmbeddingModelSpec):
     # TODO: cache from uri
-    import huggingface_hub
-    from modelscope.hub.snapshot_download import snapshot_download
+    from huggingface_hub import snapshot_download as hf_download
+    from modelscope.hub.snapshot_download import snapshot_download as ms_download
 
-    from ..llm.llm_family import symlink_local_file
+    from ..utils import retry_download, symlink_local_file
 
     cache_dir = os.path.realpath(
         os.path.join(XINFERENCE_CACHE_DIR, model_spec.model_name)
@@ -56,47 +54,29 @@ def cache(model_spec: EmbeddingModelSpec):
 
     from_modelscope: bool = model_spec.model_id.startswith("Xorbits/")
     if from_modelscope:
-        for current_attempt in range(1, MAX_ATTEMPTS + 1):
-            try:
-                download_dir = snapshot_download(
-                    model_spec.model_id, revision=model_spec.model_revision
-                )
-                break
-            except:
-                remaining_attempts = MAX_ATTEMPTS - current_attempt
-                logger.warning(
-                    f"Attempt {current_attempt} failed. Remaining attempts: {remaining_attempts}"
-                )
-        else:
-            raise RuntimeError(
-                f"Failed to download model '{model_spec.model_name}' from ModelScope "
-                f"after multiple retries"
-            )
+        download_dir = retry_download(
+            ms_download,
+            model_spec.model_name,
+            None,
+            model_spec.model_id,
+            revision=model_spec.model_revision,
+        )
         for subdir, dirs, files in os.walk(download_dir):
             for file in files:
                 relpath = os.path.relpath(os.path.join(subdir, file), download_dir)
                 symlink_local_file(os.path.join(subdir, file), cache_dir, relpath)
     else:
-        for current_attempt in range(1, MAX_ATTEMPTS + 1):
-            try:
-                huggingface_hub.snapshot_download(
-                    model_spec.model_id,
-                    revision=model_spec.model_revision,
-                    local_dir=cache_dir,
-                    local_dir_use_symlinks=True,
-                )
-                with open(meta_path, "w") as f:
-                    f.write(str(datetime.datetime.now()))
-                break
-            except:
-                remaining_attempts = MAX_ATTEMPTS - current_attempt
-                logger.warning(
-                    f"Attempt {current_attempt} failed. Remaining attempts: {remaining_attempts}"
-                )
-        else:
-            raise RuntimeError(
-                f"Failed to download model '{model_spec.model_name}' after {MAX_ATTEMPTS} attempts"
-            )
+        retry_download(
+            hf_download,
+            model_spec.model_name,
+            None,
+            model_spec.model_id,
+            revision=model_spec.model_revision,
+            local_dir=cache_dir,
+            local_dir_use_symlinks=True,
+        )
+        with open(meta_path, "w") as f:
+            f.write(str(datetime.datetime.now()))
     return cache_dir
 
 
@@ -296,7 +276,7 @@ class EmbeddingModelDescription(ModelDescription):
 
 
 def match_embedding(model_name: str) -> EmbeddingModelSpec:
-    from ..llm.llm_family import download_from_modelscope
+    from ..utils import download_from_modelscope
     from . import BUILTIN_EMBEDDING_MODELS, MODELSCOPE_EMBEDDING_MODELS
 
     if download_from_modelscope():
