+++ xinference/model/llm/llm_family.py
@@ -16,23 +16,18 @@ import logging
 import os
 import platform
 import shutil
-from pathlib import Path
 from threading import Lock
-from typing import Callable, List, Optional, Tuple, Type, Union
+from typing import List, Optional, Tuple, Type, Union
 
 from pydantic import BaseModel, Field
 from typing_extensions import Annotated, Literal
 
-from ...constants import (
-    XINFERENCE_CACHE_DIR,
-    XINFERENCE_ENV_MODEL_SRC,
-    XINFERENCE_MODEL_DIR,
-)
+from ...constants import XINFERENCE_CACHE_DIR, XINFERENCE_MODEL_DIR
+from ..utils import download_from_modelscope, retry_download, symlink_local_file
 from . import LLM
 
 logger = logging.getLogger(__name__)
 
-MAX_ATTEMPTS = 3
 DEFAULT_CONTEXT_LENGTH = 2048
 
 
@@ -94,25 +89,6 @@ BUILTIN_MODELSCOPE_LLM_FAMILIES: List["LLMFamilyV1"] = []
 UD_LLM_FAMILIES: List["LLMFamilyV1"] = []
 
 UD_LLM_FAMILIES_LOCK = Lock()
-
-
-def is_locale_chinese_simplified() -> bool:
-    import locale
-
-    try:
-        lang, _ = locale.getdefaultlocale()
-        return lang == "zh_CN"
-    except:
-        return False
-
-
-def download_from_modelscope() -> bool:
-    if os.environ.get(XINFERENCE_ENV_MODEL_SRC) == "modelscope":
-        return True
-    elif is_locale_chinese_simplified():
-        return True
-    else:
-        return False
 
 
 def download_from_self_hosted_storage() -> bool:
@@ -400,58 +376,6 @@ def _get_cache_dir(
     return cache_dir
 
 
-def symlink_local_file(path: str, local_dir: str, relpath: str) -> str:
-    from huggingface_hub.file_download import _create_symlink
-
-    # cross platform transcription of filename, to be used as a local file path.
-    relative_filename = os.path.join(*relpath.split("/"))
-    if os.name == "nt":
-        if relative_filename.startswith("..\\") or "\\..\\" in relative_filename:
-            raise ValueError(
-                f"Invalid filename: cannot handle filename '{relative_filename}' on Windows. Please ask the repository"
-                " owner to rename this file."
-            )
-    # Using `os.path.abspath` instead of `Path.resolve()` to avoid resolving symlinks
-    local_dir_filepath = os.path.join(local_dir, relative_filename)
-    if (
-        Path(os.path.abspath(local_dir))
-        not in Path(os.path.abspath(local_dir_filepath)).parents
-    ):
-        raise ValueError(
-            f"Cannot copy file '{relative_filename}' to local dir '{local_dir}': file would not be in the local"
-            " directory."
-        )
-
-    os.makedirs(os.path.dirname(local_dir_filepath), exist_ok=True)
-    real_blob_path = os.path.realpath(path)
-    _create_symlink(real_blob_path, local_dir_filepath, new_blob=False)
-    return local_dir_filepath
-
-
-def retry_download(
-    download_func: Callable,
-    llm_family: LLMFamilyV1,
-    llm_spec: "LLMSpecV1",
-    *args,
-    **kwargs,
-):
-    for current_attempt in range(1, MAX_ATTEMPTS + 1):
-        try:
-            return download_func(*args, **kwargs)
-        except:
-            remaining_attempts = MAX_ATTEMPTS - current_attempt
-            logger.warning(
-                f"Attempt {current_attempt} failed. Remaining attempts: {remaining_attempts}"
-            )
-
-    else:
-        raise RuntimeError(
-            f"Failed to download model '{llm_family.model_name}' "
-            f"(size: {llm_spec.model_size_in_billions}, format: {llm_spec.model_format}) "
-            f"after multiple retries"
-        )
-
-
 def _get_meta_path(
     cache_dir: str,
     model_format: str,
@@ -545,8 +469,11 @@ def cache_from_modelscope(
     if llm_spec.model_format == "pytorch":
         download_dir = retry_download(
             snapshot_download,
-            llm_family,
-            llm_spec,
+            llm_family.model_name,
+            {
+                "model_size": llm_spec.model_size_in_billions,
+                "model_format": llm_spec.model_format,
+            },
             llm_spec.model_id,
             revision=llm_spec.model_revision,
         )
@@ -559,8 +486,11 @@ def cache_from_modelscope(
         filename = llm_spec.model_file_name_template.format(quantization=quantization)
         download_path = retry_download(
             model_file_download,
-            llm_family,
-            llm_spec,
+            llm_family.model_name,
+            {
+                "model_size": llm_spec.model_size_in_billions,
+                "model_format": llm_spec.model_format,
+            },
             llm_spec.model_id,
             filename,
             revision=llm_spec.model_revision,
@@ -597,8 +527,11 @@ def cache_from_huggingface(
         assert isinstance(llm_spec, PytorchLLMSpecV1)
         retry_download(
             huggingface_hub.snapshot_download,
-            llm_family,
-            llm_spec,
+            llm_family.model_name,
+            {
+                "model_size": llm_spec.model_size_in_billions,
+                "model_format": llm_spec.model_format,
+            },
             llm_spec.model_id,
             revision=llm_spec.model_revision,
             local_dir=cache_dir,
@@ -610,8 +543,11 @@ def cache_from_huggingface(
         file_name = llm_spec.model_file_name_template.format(quantization=quantization)
         retry_download(
             huggingface_hub.hf_hub_download,
-            llm_family,
-            llm_spec,
+            llm_family.model_name,
+            {
+                "model_size": llm_spec.model_size_in_billions,
+                "model_format": llm_spec.model_format,
+            },
             llm_spec.model_id,
             revision=llm_spec.model_revision,
             filename=file_name,
