+++ xinference/model/utils.py
@@ -1,0 +1,104 @@
+# Copyright 2022-2023 XProbe Inc.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+import logging
+import os
+from pathlib import Path
+from typing import Callable, Dict, Optional
+
+from ..constants import XINFERENCE_ENV_MODEL_SRC
+
+logger = logging.getLogger(__name__)
+MAX_ATTEMPTS = 3
+
+
+def is_locale_chinese_simplified() -> bool:
+    import locale
+
+    try:
+        lang, _ = locale.getdefaultlocale()
+        return lang == "zh_CN"
+    except:
+        return False
+
+
+def download_from_modelscope() -> bool:
+    if os.environ.get(XINFERENCE_ENV_MODEL_SRC) == "modelscope":
+        return True
+    elif is_locale_chinese_simplified():
+        return True
+    else:
+        return False
+
+
+def symlink_local_file(path: str, local_dir: str, relpath: str) -> str:
+    from huggingface_hub.file_download import _create_symlink
+
+    # cross-platform transcription of filename, to be used as a local file path.
+    relative_filename = os.path.join(*relpath.split("/"))
+    if os.name == "nt":
+        if relative_filename.startswith("..\\") or "\\..\\" in relative_filename:
+            raise ValueError(
+                f"Invalid filename: cannot handle filename '{relative_filename}' on Windows. Please ask the repository"
+                " owner to rename this file."
+            )
+    # Using `os.path.abspath` instead of `Path.resolve()` to avoid resolving symlinks
+    local_dir_filepath = os.path.join(local_dir, relative_filename)
+    if (
+        Path(os.path.abspath(local_dir))
+        not in Path(os.path.abspath(local_dir_filepath)).parents
+    ):
+        raise ValueError(
+            f"Cannot copy file '{relative_filename}' to local dir '{local_dir}': file would not be in the local"
+            " directory."
+        )
+
+    os.makedirs(os.path.dirname(local_dir_filepath), exist_ok=True)
+    real_blob_path = os.path.realpath(path)
+    _create_symlink(real_blob_path, local_dir_filepath, new_blob=False)
+    return local_dir_filepath
+
+
+def retry_download(
+    download_func: Callable,
+    model_name: str,
+    model_info: Optional[Dict],
+    *args,
+    **kwargs,
+):
+    for current_attempt in range(1, MAX_ATTEMPTS + 1):
+        try:
+            return download_func(*args, **kwargs)
+        except:
+            remaining_attempts = MAX_ATTEMPTS - current_attempt
+            logger.warning(
+                f"Attempt {current_attempt} failed. Remaining attempts: {remaining_attempts}"
+            )
+
+    else:
+        model_size = (
+            model_info.pop("model_size", None) if model_info is not None else None
+        )
+        model_format = (
+            model_info.pop("model_format", None) if model_info is not None else None
+        )
+        if model_size is not None or model_format is not None:  # LLM models
+            raise RuntimeError(
+                f"Failed to download model '{model_name}' "
+                f"(size: {model_size}, format: {model_format}) "
+                f"after multiple retries"
+            )
+        else:  # Embedding models
+            raise RuntimeError(
+                f"Failed to download model '{model_name}' after multiple retries"
+            )
